Fill Stomach on time(U2jS2)
Our proposal is to develop an app which connects the food donator and hunger person which helps to fill the stomach of the hunger persons. It save the time of food supplier person. It stops the wastage food. User can play role in saving food wastage and help the needy.
For example: In marriage the bride groom father said to the chef nearly 300 members were visit the marriage cook the food to 300 members. After cooking food rain started unfortunately . Only 200 members were visited the marriage. Food is wastage so that bride groom father decided to donate the food to orphanages but the orphanage people were eaten food or else not eaten food by using this app we know the situation of the orphanage people.